+++ apps/test-suite/tests/FileSystemNext.ts
@@ -609,7 +609,48 @@ export async function test({ describe, expect, it, ...t }) {
           expect(file.uri).toBe(FS.documentDirectory + 'file.txt');
         });
       });
+      describe('When getting file info', () => {
+        it('executes correctly', async () => {
+          const url = `${testDirectory}execute_correctly.txt`;
+          const src = new File(url);
+          src.create();
+          src.write('Hello World');
+          const result = await src.getInfoAsync({ md5: true });
+          expect(result.exists).toBe(true);
+          if (result.exists) {
+            const { uri, size, isDirectory, modificationTime, md5 } = result;
+            expect(isDirectory).toBe(false);
+            expect(modificationTime).not.toBeNull();
+            expect(md5).not.toBeNull();
+            expect(uri).toBe(url);
+            expect(size).toBe(11);
+          }
+        });
+        it('executes correctly if options are undefined', async () => {
+          const url = `${testDirectory}executes_correctly_if_options_are_undefined.txt`;
+          const src = new File(url);
+          src.write('Hello World');
+          const result = await src.getInfoAsync();
+          expect(result.exists).toBe(true);
+          if (result.exists) {
+            expect(result.md5).toBeNull();
+          }
+        });
+        it('throws an error if file not exists', async () => {
+          const url = `${testDirectory}throws_an_error_if_file_not_exists.txt`;
+          const src = new File(url);
+          src.write('Hello world');
+          src.delete();
 
+          expect(src.exists).toBe(false);
+          try {
+            await src.getInfoAsync();
+            fail('Promise should have rejected');
+          } catch (err) {
+            expect(err).not.toBeNull();
+          }
+        });
+      });
       addAppleAppGroupsTestSuiteAsync({ describe, expect, it, ...t });
     }
   });