+++ packages/expo-file-system/src/next/ExpoFileSystem.types.ts
@@ -1,3 +1,5 @@
+import { InfoOptions } from '../FileSystem.types';
+
 export type CreateOptions = {
   /**
    * Whether to create intermediate directories if they do not exist.
@@ -139,6 +141,11 @@ export declare class File {
   delete(): void;
 
   /**
+   *  Retrieves a file info.
+   */
+  getInfoAsync(options?: InfoOptions): Promise<FileInfo>;
+
+  /**
    * A boolean representing if a file exists. `true` if the file exists, `false` otherwise.
    * Also, `false` if the application does not have read access to the file.
    */
@@ -225,3 +232,42 @@ export declare class FileHandle {
    */
   size: number | null;
 }
+
+export type FileInfo =
+  /**
+   * Object returned when file exist.
+   */
+  | {
+      /**
+       * Signifies that the requested file exist.
+       */
+      exists: true;
+      /**
+       * A `file://` URI pointing to the file. This is the same as the `fileUri` input parameter.
+       */
+      uri: string;
+      /**
+       * The size of the file in bytes. If operating on a source such as an iCloud file, only present if the `size` option was truthy.
+       */
+      size: number;
+      /**
+       * Boolean set to `true` if this is a directory and `false` if it is a file.
+       */
+      isDirectory: boolean;
+      /**
+       * The last modification time of the file expressed in seconds since epoch.
+       */
+      modificationTime: number;
+      /**
+       * Present if the `md5` option was truthy. Contains the MD5 hash of the file.
+       */
+      md5?: string;
+    }
+  /**
+   * Object returned when file do not exist.
+   */
+  | {
+      exists: false;
+      uri: string;
+      isDirectory: false;
+    };